Write lowering file test
